core-api/requirements.txt
click==8.0.3
croniter==1.0.15
ecdsa==0.17.0
Flask==2.0.2
Flask-RQ2==18.3
gunicorn==20.1.0
healthpy==1.14.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-jose==3.3.0
redis==3.5.3
rq==1.10.0
rq-scheduler==0.11.0
rsa==4.7.2
six==1.16.0
Werkzeug==2.0.2
